requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
Django==4.2.5
django-cors-headers==4.2.0
django-jazzmin==2.5.0
django-modeltranslation==0.18.11
django-payments==2.0.0
django-payments-razorpay==1.0
django-phonenumber-field==7.1.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
phonenumberslite==8.13.21
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
razorpay==1.4.1
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
